+++ changelog.md
@@ -1,6 +1,7 @@
 ### 08.04.2023, Sat
 
 - 00:15 - Endpoints `tags/` and `folders/` added.
+- 01:00 - Frontend: added list of user-created folders and tags in sidebar.
 
 ### 07.04.2023, Fr